requirements.txt
django==5.2.8
mssql-django==1.6
pyodbc==5.3.0
python-escpos==3.1
pywin32
pyusb==1.3.1
openpyxl==3.1.5
